+++ examples/exam_vector_new.jl
@@ -3,18 +3,18 @@ using CANBus
 function main()
     # bustype = CAN_20
     bustype = CAN_FD
-    vendor = KVASER
     # vendor = VECTOR
-    # vendor = SLCAN
+    # vendor = KVASER
+    vendor = SLCAN
     # vendor = SOCKETCAN
     # ch0 = "vcan0"
     # ch1 = "vcan1"
-    # ch0 = "COM3"
-    # ch1 = "COM4"
+    ch0 = "COM3"
+    ch1 = "COM4"
     # ch0 = "/dev/ttyACM0"
     # ch1 = "/dev/ttyACM1"
-    ch0 = 0
-    ch1 = 1
+    # ch0 = 0
+    # ch1 = 1
 
     ifcfg1 = InterfaceConfig(vendor, ch0, bustype, 500000)
     ifcfg1.datarate = 2000000
@@ -27,6 +27,8 @@ function main()
     ifcfg2.vendor_specific = Dict([:appname => "NewApp"])
 
     iface2 = Interface(ifcfg2)
+
+    println(iface1)
 
     frm1 = Frame(0x02, [00, 01, 02, 03, 04, 05])
     frm2 = bustype == CAN_FD ? FDFrame(0x02, collect(1:12); is_extended=true) : frm1